Module aws_sdk_machinelearning::error

Modules

idempotent_parameter_mismatch_exception

See IdempotentParameterMismatchException

internal_server_exception

See InternalServerException

invalid_input_exception

See InvalidInputException

invalid_tag_exception

See InvalidTagException

limit_exceeded_exception

See LimitExceededException

predictor_not_mounted_exception

See PredictorNotMountedException

resource_not_found_exception

See ResourceNotFoundException

tag_limit_exceeded_exception

See TagLimitExceededException

Structs

AddTagsError

CreateBatchPredictionError

CreateDataSourceFromRDSError

CreateDataSourceFromRedshiftError

CreateDataSourceFromS3Error

CreateEvaluationError

CreateMLModelError

CreateRealtimeEndpointError

DeleteBatchPredictionError

DeleteDataSourceError

DeleteEvaluationError

DeleteMLModelError

DeleteRealtimeEndpointError

DeleteTagsError

DescribeBatchPredictionsError

DescribeDataSourcesError

DescribeEvaluationsError

DescribeMLModelsError

DescribeTagsError

GetBatchPredictionError

GetDataSourceError

GetEvaluationError

GetMLModelError

IdempotentParameterMismatchException

A second request to use or change an object was not allowed. This can result from retrying a request using a parameter that was not present in the original request.

InternalServerException

An error on the server occurred when trying to process a request.

InvalidInputException

An error on the client occurred. Typically, the cause is an invalid input value.

InvalidTagException

LimitExceededException

The subscriber exceeded the maximum number of operations. This exception can occur when listing objects such as DataSource.

PredictError

PredictorNotMountedException

The exception is thrown when a predict request is made to an unmounted MLModel.

ResourceNotFoundException

A specified resource cannot be located.

TagLimitExceededException

UpdateBatchPredictionError

UpdateDataSourceError

UpdateEvaluationError

UpdateMLModelError

Enums

AddTagsErrorKind

CreateBatchPredictionErrorKind

CreateDataSourceFromRDSErrorKind

CreateDataSourceFromRedshiftErrorKind

CreateDataSourceFromS3ErrorKind

CreateEvaluationErrorKind

CreateMLModelErrorKind

CreateRealtimeEndpointErrorKind

DeleteBatchPredictionErrorKind

DeleteDataSourceErrorKind

DeleteEvaluationErrorKind

DeleteMLModelErrorKind

DeleteRealtimeEndpointErrorKind

DeleteTagsErrorKind

DescribeBatchPredictionsErrorKind

DescribeDataSourcesErrorKind

DescribeEvaluationsErrorKind

DescribeMLModelsErrorKind

DescribeTagsErrorKind

GetBatchPredictionErrorKind

GetDataSourceErrorKind

GetEvaluationErrorKind

GetMLModelErrorKind

PredictErrorKind

UpdateBatchPredictionErrorKind

UpdateDataSourceErrorKind

UpdateEvaluationErrorKind

UpdateMLModelErrorKind