Module aws_sdk_machinelearning::model

Modules

batch_prediction

See BatchPrediction

data_source

See DataSource

evaluation

See Evaluation

ml_model

See MlModel

performance_metrics

See PerformanceMetrics

prediction

See Prediction

rds_data_spec

See RdsDataSpec

rds_database

See RdsDatabase

rds_database_credentials

See RdsDatabaseCredentials

rds_metadata

See RdsMetadata

realtime_endpoint_info

See RealtimeEndpointInfo

redshift_data_spec

See RedshiftDataSpec

redshift_database

See RedshiftDatabase

redshift_database_credentials

See RedshiftDatabaseCredentials

redshift_metadata

See RedshiftMetadata

s3_data_spec

See S3DataSpec

tag

See Tag

Structs

BatchPrediction

Represents the output of a GetBatchPrediction operation.

DataSource

Represents the output of the GetDataSource operation.

Evaluation

Represents the output of GetEvaluation operation.

MlModel

Represents the output of a GetMLModel operation.

PerformanceMetrics

Measurements of how well the MLModel performed on known observations. One of the following metrics is returned, based on the type of the MLModel:

Prediction

The output from a Predict operation:

RdsDataSpec

The data specification of an Amazon Relational Database Service (Amazon RDS) DataSource.

RdsDatabase

The database details of an Amazon RDS database.

RdsDatabaseCredentials

The database credentials to connect to a database on an RDS DB instance.

RdsMetadata

The datasource details that are specific to Amazon RDS.

RealtimeEndpointInfo

Describes the real-time endpoint information for an MLModel.

RedshiftDataSpec

Describes the data specification of an Amazon Redshift DataSource.

RedshiftDatabase

Describes the database details required to connect to an Amazon Redshift database.

RedshiftDatabaseCredentials

Describes the database credentials for connecting to a database on an Amazon Redshift cluster.

RedshiftMetadata

Describes the DataSource details specific to Amazon Redshift.

S3DataSpec

Describes the data specification of a DataSource.

Tag

A custom key-value pair associated with an ML object, such as an ML model.

Enums

Algorithm

The function used to train an MLModel. Training choices supported by Amazon ML include the following:

BatchPredictionFilterVariable

A list of the variables to use in searching or filtering BatchPrediction.

DataSourceFilterVariable

A list of the variables to use in searching or filtering DataSource.

DetailsAttributes

Contains the key values of DetailsMap:

EntityStatus

Object status with the following possible values:

EvaluationFilterVariable

A list of the variables to use in searching or filtering Evaluation.

MlModelFilterVariable

MlModelType

RealtimeEndpointStatus

SortOrder

The sort order specified in a listing condition. Possible values include the following:

TaggableResourceType