Module aws_sdk_machinelearning::operation

Structs

AddTags

Adds one or more tags to an object, up to a limit of 10. Each tag consists of a key

CreateBatchPrediction

Generates predictions for a group of observations. The observations to process exist in one or more data files referenced

CreateDataSourceFromRDS

Creates a DataSource object from an Amazon Relational Database Service (Amazon RDS). A DataSource references data that can be used to perform CreateMLModel, CreateEvaluation, or CreateBatchPrediction operations.

CreateDataSourceFromRedshift

Creates a DataSource from a database hosted on an Amazon Redshift cluster. A

CreateDataSourceFromS3

Creates a DataSource object. A DataSource references data that

CreateEvaluation

Creates a new Evaluation of an MLModel. An MLModel is evaluated on a set of observations associated to a DataSource. Like a DataSource

CreateMLModel

Creates a new MLModel using the DataSource and the recipe as

CreateRealtimeEndpoint

Creates a real-time endpoint for the MLModel. The endpoint contains the URI of the MLModel; that is, the location to send real-time prediction requests for the specified MLModel.

DeleteBatchPrediction

Assigns the DELETED status to a BatchPrediction, rendering it unusable.

DeleteDataSource

Assigns the DELETED status to a DataSource, rendering it unusable.

DeleteEvaluation

Assigns the DELETED status to an Evaluation, rendering it unusable.

DeleteMLModel

Assigns the DELETED status to an MLModel, rendering it unusable.

DeleteRealtimeEndpoint

Deletes a real time endpoint of an MLModel.

DeleteTags

Deletes the specified tags associated with an ML object. After this operation is complete, you can't recover deleted tags.

DescribeBatchPredictions

Returns a list of BatchPrediction operations that match the search criteria in the request.

DescribeDataSources

Returns a list of DataSource that match the search criteria in the request.

DescribeEvaluations

Returns a list of DescribeEvaluations that match the search criteria in the request.

DescribeMLModels

Returns a list of MLModel that match the search criteria in the request.

DescribeTags

Describes one or more of the tags for your Amazon ML object.

GetBatchPrediction

Returns a BatchPrediction that includes detailed metadata, status, and data file information for a

GetDataSource

Returns a DataSource that includes metadata and data file information, as well as the current status of the DataSource.

GetEvaluation

Returns an Evaluation that includes metadata as well as the current status of the Evaluation.

GetMLModel

Returns an MLModel that includes detailed metadata, data source information, and the current status of the MLModel.

Predict

Generates a prediction for the observation using the specified ML Model.

UpdateBatchPrediction

Updates the BatchPredictionName of a BatchPrediction.

UpdateDataSource

Updates the DataSourceName of a DataSource.

UpdateEvaluation

Updates the EvaluationName of an Evaluation.

UpdateMLModel

Updates the MLModelName and the ScoreThreshold of an MLModel.