Module aws_sdk_machinelearning::output

Modules

add_tags_output

See AddTagsOutput

create_batch_prediction_output

See CreateBatchPredictionOutput

create_data_source_from_rds_output

See CreateDataSourceFromRdsOutput

create_data_source_from_redshift_output

See CreateDataSourceFromRedshiftOutput

create_data_source_from_s3_output

See CreateDataSourceFromS3Output

create_evaluation_output

See CreateEvaluationOutput

create_ml_model_output

See CreateMlModelOutput

create_realtime_endpoint_output

See CreateRealtimeEndpointOutput

delete_batch_prediction_output

See DeleteBatchPredictionOutput

delete_data_source_output

See DeleteDataSourceOutput

delete_evaluation_output

See DeleteEvaluationOutput

delete_ml_model_output

See DeleteMlModelOutput

delete_realtime_endpoint_output

See DeleteRealtimeEndpointOutput

delete_tags_output

See DeleteTagsOutput

describe_batch_predictions_output

See DescribeBatchPredictionsOutput

describe_data_sources_output

See DescribeDataSourcesOutput

describe_evaluations_output

See DescribeEvaluationsOutput

describe_ml_models_output

See DescribeMlModelsOutput

describe_tags_output

See DescribeTagsOutput

get_batch_prediction_output

See GetBatchPredictionOutput

get_data_source_output

See GetDataSourceOutput

get_evaluation_output

See GetEvaluationOutput

get_ml_model_output

See GetMlModelOutput

predict_output

See PredictOutput

update_batch_prediction_output

See UpdateBatchPredictionOutput

update_data_source_output

See UpdateDataSourceOutput

update_evaluation_output

See UpdateEvaluationOutput

update_ml_model_output

See UpdateMlModelOutput

Structs

AddTagsOutput

Amazon ML returns the following elements.

CreateBatchPredictionOutput

Represents the output of a CreateBatchPrediction operation, and is an acknowledgement that Amazon ML received the request.

CreateDataSourceFromRdsOutput

Represents the output of a CreateDataSourceFromRDS operation, and is an acknowledgement that Amazon ML received the request.

CreateDataSourceFromRedshiftOutput

Represents the output of a CreateDataSourceFromRedshift operation, and is an acknowledgement that Amazon ML received the request.

CreateDataSourceFromS3Output

Represents the output of a CreateDataSourceFromS3 operation, and is an acknowledgement that Amazon ML received the request.

CreateEvaluationOutput

Represents the output of a CreateEvaluation operation, and is an acknowledgement that Amazon ML received the request.

CreateMlModelOutput

Represents the output of a CreateMLModel operation, and is an acknowledgement that Amazon ML received the request.

CreateRealtimeEndpointOutput

Represents the output of an CreateRealtimeEndpoint operation.

DeleteBatchPredictionOutput

Represents the output of a DeleteBatchPrediction operation.

DeleteDataSourceOutput

Represents the output of a DeleteDataSource operation.

DeleteEvaluationOutput

Represents the output of a DeleteEvaluation operation. The output indicates that Amazon Machine Learning (Amazon ML) received the request.

DeleteMlModelOutput

Represents the output of a DeleteMLModel operation.

DeleteRealtimeEndpointOutput

Represents the output of an DeleteRealtimeEndpoint operation.

DeleteTagsOutput

Amazon ML returns the following elements.

DescribeBatchPredictionsOutput

Represents the output of a DescribeBatchPredictions operation. The content is essentially a list of BatchPredictions.

DescribeDataSourcesOutput

Represents the query results from a DescribeDataSources operation. The content is essentially a list of DataSource.

DescribeEvaluationsOutput

Represents the query results from a DescribeEvaluations operation. The content is essentially a list of Evaluation.

DescribeMlModelsOutput

Represents the output of a DescribeMLModels operation. The content is essentially a list of MLModel.

DescribeTagsOutput

Amazon ML returns the following elements.

GetBatchPredictionOutput

Represents the output of a GetBatchPrediction operation and describes a BatchPrediction.

GetDataSourceOutput

Represents the output of a GetDataSource operation and describes a DataSource.

GetEvaluationOutput

Represents the output of a GetEvaluation operation and describes an Evaluation.

GetMlModelOutput

Represents the output of a GetMLModel operation, and provides detailed information about a MLModel.

PredictOutput

UpdateBatchPredictionOutput

Represents the output of an UpdateBatchPrediction operation.

UpdateDataSourceOutput

Represents the output of an UpdateDataSource operation.

UpdateEvaluationOutput

Represents the output of an UpdateEvaluation operation.

UpdateMlModelOutput

Represents the output of an UpdateMLModel operation.